chore(blask): Rename modules with snake_case names
