+++ blask/__init__.py
@@ -17,4 +17,4 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 """
 
-from blask.blaskapp import BlaskApp
+from blask.blask_app import BlaskApp
