+++ tests/settings_test.py
@@ -4,8 +4,8 @@ from pathlib import Path
 import pytest
 from dotenv import load_dotenv
 
-from blask import blasksettings
-from blask.blasksettings import BlaskSettings
+from blask import blask_settings
+from blask.blask_settings import BlaskSettings
 
 
 class TestBlaskSettings:
@@ -14,14 +14,14 @@ class TestBlaskSettings:
 
         # remove all environment variables
         removed_vars = {}
-        for key in blasksettings.DEFAULT_SETTINGS:
+        for key in blask_settings.DEFAULT_SETTINGS:
             if key in os.environ:
                 value = os.environ.pop(key)
                 removed_vars[key] = value
 
         settings = BlaskSettings()
-        for kw in blasksettings.DEFAULT_SETTINGS.keys():
-            assert settings[kw] == blasksettings.DEFAULT_SETTINGS[kw]
+        for kw in blask_settings.DEFAULT_SETTINGS.keys():
+            assert settings[kw] == blask_settings.DEFAULT_SETTINGS[kw]
 
         # put environment variables back
         for key, value in removed_vars.items():
@@ -31,7 +31,7 @@ class TestBlaskSettings:
 
         # remove all environment variables
         removed_vars = {}
-        for key in blasksettings.DEFAULT_SETTINGS:
+        for key in blask_settings.DEFAULT_SETTINGS:
             if key in os.environ:
                 value = os.environ.pop(key)
                 removed_vars[key] = value
@@ -40,13 +40,13 @@ class TestBlaskSettings:
         os.environ['BLASK_SETTINGS'] = 'tests.testsettings'
 
         settings = BlaskSettings()
-        for kw in blasksettings.DEFAULT_SETTINGS.keys():
+        for kw in blask_settings.DEFAULT_SETTINGS.keys():
             if kw == 'post_dir':
                 assert settings[kw] == str(Path('.').resolve() / 'posts2')
             elif kw == 'title':
                 assert settings[kw] == 'The mantis revenge!'
             else:
-                assert settings[kw] == blasksettings.DEFAULT_SETTINGS[kw]
+                assert settings[kw] == blask_settings.DEFAULT_SETTINGS[kw]
         del(os.environ['BLASK_SETTINGS'])
 
         # put environment variables back
@@ -57,7 +57,7 @@ class TestBlaskSettings:
     def test_from_dotenv(self):
         # remove all environment variables
         removed_vars = {}
-        for key in blasksettings.DEFAULT_SETTINGS:
+        for key in blask_settings.DEFAULT_SETTINGS:
             if key in os.environ:
                 value = os.environ.pop(key)
                 removed_vars[key] = value
@@ -67,16 +67,16 @@ class TestBlaskSettings:
 
         # changed for new pytest
         settings = BlaskSettings()
-        for kw in blasksettings.DEFAULT_SETTINGS.keys():
+        for kw in blask_settings.DEFAULT_SETTINGS.keys():
             if kw == 'post_dir':
                 assert settings[kw] == 'posts-env'
             elif kw == 'title':
                 assert settings[kw] == "The mantis with dotenv?!"
             else:
-                assert settings[kw] == blasksettings.DEFAULT_SETTINGS[kw]
+                assert settings[kw] == blask_settings.DEFAULT_SETTINGS[kw]
 
         # remove all test environment variables
-        for key in blasksettings.DEFAULT_SETTINGS:
+        for key in blask_settings.DEFAULT_SETTINGS:
             if key in os.environ:
                 os.environ.pop(key)
 
@@ -88,7 +88,7 @@ class TestBlaskSettings:
 
         # remove all environment variables
         removed_vars = {}
-        for key in blasksettings.DEFAULT_SETTINGS:
+        for key in blask_settings.DEFAULT_SETTINGS:
             if key in os.environ:
                 value = os.environ.pop(key)
                 removed_vars[key] = value
@@ -98,13 +98,13 @@ class TestBlaskSettings:
         }
         settings = BlaskSettings(**kwsettings)
 
-        for kw in blasksettings.DEFAULT_SETTINGS.keys():
+        for kw in blask_settings.DEFAULT_SETTINGS.keys():
             if kw == 'post_dir':
                 assert settings[kw] == str(Path.cwd() / 'mantispost_dir')
             elif kw == 'title':
                 assert settings[kw] == 'The mantis has you!'
             else:
-                assert settings[kw] == blasksettings.DEFAULT_SETTINGS[kw]
+                assert settings[kw] == blask_settings.DEFAULT_SETTINGS[kw]
 
         # put original environment variables back
         for key, value in removed_vars.items():
